+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# 0.26.2
+## Features Added
+  - `Device::create` allows users to provide metadata and tags at creation time
+  - `Device::update` and convenience methods allow metadata and tags to be updated after creation
+  - Token APIs allow you to specify a set of API scopes to limit the token's permissions
+
 # 0.26.1
 ## Bugs Fixed
   - Added a docstring for `CreateDatasetIfNotExistsRequest` to fix rendering issue in docs